Refactor and installing of additional deps
