+++ package.json
@@ -24,9 +24,11 @@
     "nodemon": "^3.1.4",
     "react": "18.3.1",
     "react-dom": "18.3.1",
+    "react-hot-toast": "^2.4.1",
     "react-router-dom": "^6.26.2",
     "recoil": "^0.7.7",
-    "tslib": "^2.3.0"
+    "tslib": "^2.3.0",
+    "uuid": "^10.0.0"
   },
   "devDependencies": {
     "@nrwl/express": "^19.7.3",
